LayoutTests/editing/execCommand/create-list-from-range-selection-expected.txt


<DIV id="test" contentEditable="true">
<OL>
<LI>
<#text>asd</#text>
<SPAN id="start">
<#text>fo<selection-anchor>o</#text>
</SPAN>
</LI>
<LI>
<#text>bar</#text>
</LI>
<LI>
<#text>baz</#text>
</LI>
<LI>
<#text>foo</#text>
</LI>
<LI>
<#text>bar</#text>
</LI>
<LI>
<#text>ba<selection-focus>z</#text>
</LI>
</OL>
</DIV>
